Split out witchcraft-jwt
