+++ witchcraft-server/src/service/witchcraft_mdc.rs
@@ -14,7 +14,7 @@
 
 use crate::logging;
 use crate::service::request_id::RequestId;
-use crate::service::unverified_jwt::UnverifiedJwt;
+use witchcraft_jwt::unverified_jwt::UnverifiedJwt;
 use crate::service::{Layer, Service};
 use http::Request;
 use witchcraft_log::mdc;
